Fix circular dependencies import
